+++ js/modules/state.js
@@ -1,0 +1,7 @@
+export const state = {
+  currentUser: null,
+  currentLocation: { lat: null, lon: null, placeId: null },
+  currentNoteId: null,
+  canvas: null,
+  ctx: null,
+};
